rename ntlm_parser to negotiate_parser, add TDS LOGIN7 helpers
