+++ nxc/protocols/mssql.py
@@ -9,7 +9,7 @@ from nxc.connection import requires_admin
 from nxc.helpers.misc import gen_random_string
 from nxc.logger import NXCAdapter
 from nxc.helpers.bloodhound import add_user_bh
-from nxc.helpers.ntlm_parser import parse_challenge
+from nxc.helpers.negotiate_parser import parse_challenge, login7_integrated_auth_error_message
 from nxc.helpers.powershell import create_ps_command
 from nxc.protocols.mssql.mssqlexec import MSSQLEXEC
 
@@ -78,31 +78,6 @@ class mssql(connection):
             return func(self, *args, **kwargs)
         return wrapper
 
-    @staticmethod
-    def decode_tds_info_error_msgtext(data, offset):
-        remaining = len(data) - offset
-        if remaining < 3 or data[offset] not in (TDS_ERROR_TOKEN, TDS_INFO_TOKEN):
-            return None
-        payload_len = int.from_bytes(data[offset + 1 : offset + 3], "little")
-        if payload_len < 8 or remaining < 3 + payload_len:
-            return None
-        try:
-            token = tds.TDS_INFO_ERROR(data[offset:])
-            text = token["MsgText"].decode("utf-16le").strip()
-        except Exception:
-            return None
-        return text or None
-
-    @staticmethod
-    def login7_integrated_auth_error_message(packet_data, data_after_login_header):
-        token_markers = (TDS_ERROR_TOKEN, TDS_INFO_TOKEN)
-        for buf in filter(None, (packet_data, data_after_login_header)):
-            for offset in (i for i in range(len(buf)) if buf[i] in token_markers):
-                msg = mssql.decode_tds_info_error_msgtext(buf, offset)
-                if msg:
-                    return msg
-        return None
-
     def check_if_admin(self):
         self.admin_privs = False
         try:
@@ -162,7 +137,7 @@ class mssql(connection):
             return False
         else:
             if not challenge.startswith(b"NTLMSSP\x00"):
-                error_msg = self.login7_integrated_auth_error_message(login_response, challenge)
+                error_msg = login7_integrated_auth_error_message(login_response, challenge)
                 detail = f": {error_msg}" if error_msg else ""
                 self.logger.fail(f"Server does not support Integrated Windows Authentication{detail}")
             else:
