+++ nxc/protocols/wmi.py
@@ -1,7 +1,7 @@
 import os
 from io import StringIO
 
-from nxc.helpers.ntlm_parser import parse_challenge
+from nxc.helpers.negotiate_parser import parse_challenge
 from nxc.config import process_secret
 from nxc.connection import connection, dcom_FirewallChecker, requires_admin
 from nxc.logger import NXCAdapter
